+++ library/ruler/src/test/java/io/mehow/ruler/DistanceUnitGenerator.kt
@@ -4,7 +4,7 @@ import io.kotest.property.Exhaustive
 
 internal object DistanceUnitGenerator : Exhaustive<LengthUnit<*>>() {
   override val values: List<LengthUnit<*>>
-    get() = SiLengthUnit.values + ImperialLengthUnit.values
+    get() = SiLengthUnit.units + ImperialLengthUnit.units
 
   fun createLength(distance: Distance, unit: LengthUnit<*>) = when (unit) {
     is SiLengthUnit -> distance.toLength(unit)